+++ dev/ASPNetMVC_frameworkNet/Q_Repository/Models/Categorie.cs
@@ -20,6 +20,6 @@ namespace Q_Repository.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Reference { get; set; }
-        public double ListPrice { get; set; }
+        public decimal ListPrice { get; set; }
     }
 }
